refacto: split `io.rs` and move to separate folder
