+++ zenoh-flow/src/lib.rs
@@ -39,6 +39,7 @@ use const_format::formatcp;
 
 pub use ::zenoh_flow_derive;
 
+pub mod io;
 pub mod macros;
 pub mod model;
 pub mod runtime;
@@ -52,18 +53,13 @@ pub use anyhow::anyhow;
 pub use zfresult::{DaemonResult, ZFResult as Result};
 
 pub mod prelude {
-
+    pub use crate::io::{Input, InputRaw, Inputs, Output, OutputRaw, Outputs};
     pub use crate::traits::{DowncastAny, Node, Operator, Sink, Source, ZFData};
-    pub use crate::types::{
-        Configuration, Context, Input, Inputs, Message, NodeId, Output, Outputs, Payload, PortId,
-        RuntimeId,
-    };
-    pub use crate::zferror;
-    pub use crate::{export_operator, export_sink, export_source};
-
-    pub use crate::zfresult::{Error, ErrorKind, ZFResult as Result};
-
+    pub use crate::types::{Configuration, Context, Message, NodeId, Payload, PortId, RuntimeId};
     pub use crate::zenoh_flow_derive::ZFData;
+    pub use crate::zferror;
+    pub use crate::zfresult::{Error, ErrorKind, ZFResult as Result};
+    pub use crate::{export_operator, export_sink, export_source};
 }
 
 /// Commit id of latest commit on Zenoh Flow
