+++ zenoh-flow/src/runtime/dataflow/instance/mod.rs
@@ -17,10 +17,11 @@ pub mod runners;
 use self::runners::connector::{ZenohReceiver, ZenohSender};
 use self::runners::Runner;
 use super::DataFlow;
+use crate::io::{Inputs, Outputs};
 use crate::model::record::{LinkRecord, ZFConnectorKind};
 use crate::prelude::{Context, Node};
 use crate::runtime::InstanceContext;
-use crate::types::{Inputs, NodeId, Outputs};
+use crate::types::NodeId;
 use crate::zfresult::ErrorKind;
 use crate::Result;
 use crate::{bail, zferror};
